+++ package.json
@@ -4,7 +4,7 @@
   "description": "React frontend for IPD module",
   "main": "dist/index.js",
   "author": "bahmni@thoughtworks.com",
-  "license": "AGPL-3.0",
+  "license": "MPL-2.0",
   "devDependencies": {
     "@babel/plugin-transform-runtime": "^7.22.4",
     "@babel/preset-env": "^7.22.4",
@@ -44,6 +44,8 @@
     "carbon-components": "^10.19.0",
     "carbon-components-react": "^7.25.0",
     "carbon-icons": "^7.0.7",
+    "file-loader": "^6.2.0",
+    "prop-types": "^15.8.1",
     "react": "16.14.0",
     "react-dom": "16.14.0",
     "react-router-dom": "^6.14.1"
