+++ src/App.js
@@ -1,8 +1,9 @@
 import React from "react";
-import Login from "./pages/Login";
+import Login from "./pages/Login/Login";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Navbar from "./components/commons/Navbar";
-import Footer from "./components/commons/Footer";
+import Navbar from "./components/commons/Navbar/Navbar";
+import Footer from "./components/commons/Footer/Footer";
+import "./styles/styles.scss";
 
 const App = () => {
   return (
