+++ src/assets/README.md
@@ -1,0 +1,7 @@
+# Assets Folder
+
+The images in the assets/images folder are sourced from Pexels which provides stock images that can be used for commercial purposes without any fee or attribution.
+
+You can find Pexels [here](https://www.pexels.com/) and the specific image links [here](https://www.pexels.com/photo/a-corridor-of-a-hospital-5203594/) and [here](https://www.pexels.com/photo/woman-smiling-while-holding-a-stethoscope-6749778/).
+
+Their attribution policy can be found [here](https://www.pexels.com/license/).
