+++ src/components/Landing/LandingSection.scss
@@ -1,0 +1,115 @@
+@import "../../styles/variables";
+
+.hero__container {
+  padding: 2%;
+  overflow-x: hidden;
+  width: 80%;
+}
+
+.hero__content {
+  background-color: #ffffff;
+  min-width: 80%;
+}
+
+.bx--grid.hero__content__mobile {
+  display: none;
+  background-color: #ffffff;
+}
+
+.hero__text {
+  padding: 6%;
+  margin-left: 2rem;
+}
+
+.hero__section > * {
+  margin-bottom: 1rem;
+}
+
+.hero__section > p {
+  margin-bottom: 0.2rem;
+  font-size: 18px;
+}
+
+.hero__section span {
+  color: $secondary_color;
+}
+
+.hero__form {
+  margin-top: 2rem;
+}
+
+.input__section {
+  text-align: center;
+}
+
+.error__text {
+  color: $error_text_color;
+  height: 1rem;
+}
+
+.hero___imagesection {
+  position: relative;
+}
+
+.hero__image {
+  height: 500px;
+  object-fit: contain;
+  position: absolute;
+  right: 0;
+  bottom: 0;
+  margin-right: 2rem;
+}
+
+.hero__image__mobile {
+  height: 350px;
+}
+
+/* ------------ Media Queries ------------- */
+
+@media screen and (max-width: 985px) {
+  .hero__text {
+    padding: 6%;
+  }
+  .hero__image {
+    height: 450px;
+    margin-right: 0.5rem;
+  }
+  .hero__text {
+    margin-left: 0.5rem;
+  }
+  .hero__section h1 {
+    font-size: 2rem;
+  }
+  .hero__section > p {
+    margin-bottom: 0.2rem;
+    font-size: 1rem;
+  }
+}
+
+@media screen and (max-width: 830px) {
+  .hero__image {
+    height: 400px;
+  }
+}
+
+@media screen and (max-width: 768px) {
+  .bx--row.hero__content {
+    display: none;
+  }
+  .bx--grid.hero__content__mobile {
+    display: block;
+  }
+  .hero___imagesection {
+    text-align: center;
+  }
+  .hero__image,
+  .hero__text {
+    margin: 0;
+  }
+}
+
+@media screen and (max-width: 480px) {
+  .hero__image__mobile {
+    height: 200px;
+  }
+}
