+++ src/components/commons/Toaster/Toaster.scss
@@ -1,0 +1,7 @@
+.toaster__notification {
+  position: fixed;
+  top: 0;
+  left: 50%;
+  transform: translateX(-50%);
+  z-index: 9999;
+}
