+++ src/index.js
@@ -2,6 +2,5 @@ import React from "react";
 import reactDom from "react-dom";
 import App from "./App";
 import "bahmni-carbon-ui/styles.css";
-import "./styles/styles.scss";
 
 reactDom.render(<App />, document.getElementById("root"));
