--- src/pages/Login.js
@@ -1,12 +1,0 @@
-import React from "react";
-import LandingSection from "../components/LandingSection";
-
-const Login = () => {
-  return (
-    <div data-testid="login-container" className="login__container">
-      <LandingSection />
-    </div>
-  );
-};
-
-export default Login;
