+++ src/pages/Login/Login.scss
@@ -1,0 +1,13 @@
+.login__container {
+  background: url("../../assets/images/background.png");
+  background-attachment: fixed;
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: center center;
+  margin-top: 3rem;
+  padding: 1rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  flex: 1;
+}
