+++ src/styles/_variables.scss
@@ -1,0 +1,6 @@
+/* Colors */
+
+$primary_color: #c6e2ff;
+$secondary_color: #ff7416;
+$link_color: #000000;
+$error_text_color: #ff0000;
