--- src/styles/common/_button.scss
@@ -1,3 +1,0 @@
-.primary__button {
-  background-color: $secondary_color !important;
-}
