--- src/styles/common/_commonIndex.scss
@@ -1,7 +1,0 @@
-@import "variables";
-@import "navbar";
-@import "footer";
-@import "login";
-@import "landing";
-@import "button";
-@import "toaster";
