--- src/styles/common/_footer.scss
@@ -1,13 +1,0 @@
-.footer__container {
-  bottom: 0;
-  position: relative;
-  width: 100%;
-  background-color: $primary_color;
-  padding: 1rem;
-}
-
-.footer__row.bx--row {
-  justify-content: center;
-  text-align: center;
-  margin-bottom: 0.5rem;
-}
