--- src/styles/common/_login.scss
@@ -1,13 +1,0 @@
-.login__container {
-  background: url(https://i.ibb.co/Bfd2d2B/Hero-bg.png);
-  background-attachment: fixed;
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-position: center center;
-  margin-top: 3rem;
-  padding: 1rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  flex: 1;
-}
