--- src/styles/common/_navbar.scss
@@ -1,40 +1,0 @@
-.header__container {
-  background-color: $primary_color !important;
-  flex-direction: row;
-  justify-content: space-between;
-  padding: 0 1rem;
-}
-
-.header__icon {
-  color: $link_color !important;
-  border: none !important;
-  padding: 0rem 1rem !important;
-}
-
-.header__item {
-  list-style: none;
-  text-decoration: none;
-  a {
-    color: $link_color !important;
-  }
-  a:hover {
-    background-color: $primary_color !important;
-  }
-  .bx--header__menu-item {
-    border: none;
-  }
-}
-
-.header__navlinks {
-  display: flex;
-  justify-content: space-between;
-  flex-direction: row;
-  align-items: center;
-  background-color: $primary_color;
-}
-
-.sidebar__icon {
-  svg {
-    fill: $link_color !important;
-  }
-}
