--- src/styles/common/_toaster.scss
@@ -1,7 +1,0 @@
-.toaster__notification {
-  position: fixed;
-  top: 0;
-  left: 50%;
-  transform: translateX(-50%);
-  z-index: 9999;
-}
