+++ src/styles/styles.scss
@@ -1,4 +1,4 @@
-@import "common/commonIndex";
+@import "variables";
 
 @import "~carbon-components/scss/globals/scss/styles.scss";
