+++ webpack.config.js
@@ -66,6 +66,18 @@ const commonConfig = {
           "sass-loader",
         ],
       },
+      {
+        test: /\.(gif|png|avif|jpe?g)$/,
+        use: [
+          {
+            loader: "file-loader",
+            options: {
+              name: "[name].[ext]",
+              outputPath: "assets/images/",
+            },
+          },
+        ],
+      },
     ],
   },
   devServer: {},
